Move NTCTSNumberData in file
