+++ src/libs/antares/CMakeLists.txt
@@ -506,7 +506,7 @@ set(SRC
     # TODO : should be in a target_include_directories : but need to remove all .h from all SRC
     Enum.hpp
     Enum.hxx
-)
+		study/scenario-builder/NTCTSNumberData.cpp study/scenario-builder/NTCTSNumberData.h)
 
 
 add_library( libantares-core-calendar
