+++ src/libs/antares/study/scenario-builder/rules.h
@@ -32,6 +32,7 @@
 #include "../fwd.h"
 #include "TSnumberData.h"
 #include "hydroLevelsData.h"
+#include "NTCTSNumberData.h"
 #include <map>
 #include <memory>
